[Rating] Move from lab to core
