+++ docs/pages/api-docs/rating.md
@@ -1,5 +1,5 @@
 ---
-filename: /packages/material-ui-lab/src/Rating/Rating.js
+filename: /packages/material-ui/src/Rating/Rating.js
 ---
 
 <!--- This documentation is automatically generated, do not try to edit it. -->
@@ -11,9 +11,9 @@ filename: /packages/material-ui-lab/src/Rating/Rating.js
 ## Import
 
 ```js
-import Rating from '@material-ui/lab/Rating';
+import Rating from '@material-ui/core/Rating';
 // or
-import { Rating } from '@material-ui/lab';
+import { Rating } from '@material-ui/core';
 ```
 
 You can learn more about the difference by [reading this guide](/guides/minimizing-bundle-size/).
@@ -76,7 +76,7 @@ You can override the style of the component thanks to one of these customization
 - With a [global class name](/customization/components/#overriding-styles-with-global-class-names).
 - With a theme and an [`overrides` property](/customization/globals/#css).
 
-If that's not sufficient, you can check the [implementation of the component](https://github.com/mui-org/material-ui/blob/next/packages/material-ui-lab/src/Rating/Rating.js) for more detail.
+If that's not sufficient, you can check the [implementation of the component](https://github.com/mui-org/material-ui/blob/next/packages/material-ui/src/Rating/Rating.js) for more detail.
 
 ## Demos
