+++ docs/src/modules/utils/getJsxPreview.test.js
@@ -7,7 +7,7 @@ describe('getJsxPreview', () => {
       getJsxPreview(
         `
 import * as React from 'react';
-import Rating from '@material-ui/lab/Rating';
+import Rating from '@material-ui/core/Rating';
 
 export default function HalfRating() {
   return <Rating name="half-rating" value={2.5} precision={0.5} />;
