+++ docs/src/pages/components/rating/RatingSize.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Rating from '@material-ui/lab/Rating';
+import Rating from '@material-ui/core/Rating';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme) => ({
