+++ docs/src/pages/components/rating/SimpleRating.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Rating from '@material-ui/lab/Rating';
+import Rating from '@material-ui/core/Rating';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
