+++ docs/src/pages/components/rating/rating.md
@@ -3,7 +3,6 @@ title: Rating React component
 components: Rating
 githubLabel: 'component: Rating'
 waiAria: https://www.w3.org/WAI/tutorials/forms/custom-controls/#a-star-rating
-packageName: '@material-ui/lab'
 ---
 
 # Rating
