+++ docs/src/pages/guides/migration-v4/migration-v4.md
@@ -524,6 +524,13 @@ const theme = createMuitheme({
 
 ### Rating
 
+- Move the component from the lab to the core. The component is now stable.
+
+  ```diff
+  -import Alert from '@material-ui/lab/Rating';
+  +import Alert from '@material-ui/core/Rating';
+  ```
+
 - Change the default empty icon to improve accessibility.
   If you have a custom `icon` prop but no `emptyIcon` prop, you can restore the previous behavior with:
