+++ packages/material-ui/src/Rating/Rating.test.js
