+++ packages/material-ui/src/Rating/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './Rating';
+export * from './Rating';
