+++ packages/material-ui/src/Rating/index.js
@@ -1,0 +1,1 @@
+export { default } from './Rating';
